Initial Git Commi
This project was initially intended as a way to practice building a
project from scratch. By using the React and Express framework, the
project was meant to be a way to practice the PERN stack. But after some
consideration, the project has evolved into a proper website that could
assist me and other later on in the future. Therefore, it makes sense to
have proper version control, even though the project is probably almost
done with it's alpha stage. Better late than never.