Minor Tab bug fix
Calling this commit a fix is absurd, but it does fix the problem of the
dates not showing up in the WalletView component. The fix gives rise to
another problem however. The component will now continously call the
backend for data that it already has. This won't be an app breaking
problem but it will cause a lot of data to be wasted. A solution for
this could be to send the dates from the server instead.