Added responsive UI features
The whole app now feels good to use on the phone. There's a slight issue
with the AddTransaction page where the textboxes appear to long in the
desktop version and there is a slight padding issue with the walletView
page. But other than that, the whole app looks crisp and responsive.