Added Buget frontend functionality and bug fixes.
The front end of the budgetting feature is implemented. Certain bugs
were fixed, but the feature still needs to be vigorously tested since it
requires a day to see if the functionality works. Minor upgrades to the
UI where certain functions are explained to the user. The app is now
ready for live testing.